refactor #10 - 서버 클라이언트와 브라우저 클라이언트 get 코드 분리
