+++ src/utils/supabase/serverClient.ts
@@ -1,0 +1,22 @@
+import { createServerClient } from '@supabase/ssr';
+import { cookies } from 'next/headers';
+import { SupabaseClient } from '@supabase/supabase-js';
+
+const SUPABASE_URL = process.env.NEXT_PUBLIC_SUPABASE_URL!;
+const SUPABASE_ANON_KEY = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!;
+
+export const getServerClient = (): SupabaseClient => {
+  const cookieStore = cookies();
+  return createServerClient(SUPABASE_URL, SUPABASE_ANON_KEY, {
+    cookies: {
+      getAll() {
+        return cookieStore.getAll();
+      },
+      setAll(cookiesToSet) {
+        try {
+          cookiesToSet.forEach(({ name, value, options }) => cookieStore.set(name, value, options));
+        } catch {}
+      }
+    }
+  });
+};
